Download functionality.
Ran into issues where I was losing lines to overflow at the bottom, narrowed it down to rendered tables wrapping their text (even though they still had plenty space) ONLY in the screenshot. So the height was static compute, but the text was wrapped, which pushed all elements down. This led to bottom lines getting pushed out of the container.

Not sure how to fix yet.